refactor: initial work-in-progress implementation of streaming
This is the first PR to update some of the extension host side to
support streaming and 'infinitely large files' as needed for DAP.

Here I focus on creating a new "hex document model" which manages edits
and allows reading bytes with those edits applied. There's some gnarly
code here, so I also got a test setup going for some sanity checking.
The main additional complexity here is that this model now supports
insertions and removals in the middle of memory, not just at the end.
Though this probably won't be used in the debug case initially, I
thought it was still useful if I was rewriting things anyway...

Under the model lies the `fileAccessor` which is a simple I/O model. The
current accessor just uses the same VS Code APIs, but we could have some
extra casing that uses the native Node APIs to support ultra large
files when running in the Node extension host.

I also formalized the communication a bit for upcoming refactors there.
I haven't touched any of the webview-side code yet, and there are still
some parts of the extension host I haven't reworked yet (find/replace
mainly)