init: edit mode and input byte insertion
Allows to manually insert bytes into the file, just like replace. To
choose between these, there is a edit mode that can be changed
either in the status bar or through a command.
To visually distinguish them inside the editor, each mode has an unique
cursor on the focusedByte. When Inserting the cursor is a text cursor
and in Replacing it is a box.