wip: added support to insert at eof
Initially was using a text cursor to append at a certain offset, but it
was terrible UX due to having to specifically click on the rhs of the offset.
So I opted to add an extra data cell at the end with a "+" icon. This way the user
can easily figure out how to append bytes and does not complicate byte insertion.